pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "inngest>=0.5.9",
    "llama-index-core>=0.14.3",
    "llama-index-readers-file>=0.5.4",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "uvicorn>=0.37.0",
]
